FIX for issues with ActiveRecord 3.1+
When using AR 3.0, empty serialized fields (like `Audit#audited_changes`) were
either not written to the DB or not serialized before writing.  In effect, an
empty serialized field was stored as `NULL`.  `nil` in the model was `NULL`
in the DB.

In AR 3.1, that changed; the field is apparently serialized no matter what,
resulting in a YAML string representing a `nil`.  Nonetheless, it's a string
and not `NULL`.  Thus, the explicit check in the `Audit.modifying` scope
failed.